requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
PyYAML
Markdown
Pygments
boto3
Pillow
bleach
tinycss2
waitress
